+++ hsb-bo/src/main/java/com/hbs/hsbbo/common/repository/CodeParentRepository.java
@@ -3,6 +3,6 @@ package com.hbs.hsbbo.common.repository;
 import com.hbs.hsbbo.common.domain.entity.CodeParent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CodeParentRepository extends JpaRepository<CodeParent, Integer>{
+public interface CodeParentRepository extends JpaRepository<CodeParent, Long>{
 
 }
